Create declarative_link_capturing.md, a related proposal
Adds a new proposal for "declarative link capturing" in the manifest.

This serves as a side proposal to launch event. It allows for many of the launch event use cases
to be done declaratively, without having to implement the full event, but has an upgrade path to
the full event should we need it.