Replace appmanifest's launch algorithm with one that has launch_handler behaviors (#79)
This builds on top of w3c/manifest#1056 which adds a launch web application algorithm to the manifest spec. This updates the launch handler spec to replace that algorithm with its own. This enables other webapp specs (not in WICG) to gain the launch_handler behaviour without directly referencing this WICG spec.

Additionally this replaces the concept of "web app launch client" with the existing "application context" concept from the manifest spec which is effectively the same thing.

#67 #69 #84
https://github.com/WICG/web-app-launch/pull/79